requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
Pillow==9.4.0
numpy==1.25.1
requests==2.31.0
flask-bootstrap==3.3.7.1
